+++ src/main/java/com/example/CSIA/api/UserController.java
@@ -24,35 +24,55 @@ public class UserController {
         this.userRepository = userRepository;
     }
 
-    // Adds a User
+    /**
+     * This method adds a user to the user database
+     * @param user Valid user entity parsed through JSON file
+     * @return Success message
+     */
     @PostMapping
     public ResponseEntity<?> insertUser(@NonNull @Valid @RequestBody User user) {
         userRepository.save(user);
         return ResponseEntity.ok("Success!");
     }
 
-    // Get All Users
+    /**
+     * This method gets all users in the database
+     * @return List of all users
+     */
     @GetMapping
     public ResponseEntity<?> getAllUser() {
         return ResponseEntity.ok(userRepository.findAll());
     }
 
-    // Get All ID
+    /**
+     * This method gets all user's ID in the database
+     * @return List of all user IDs.
+     */
     @GetMapping("/id")
-    public List<UUID> getAllUserID() {
+    public ResponseEntity<?> getAllUserID() {
         List<UUID> uuidList = new ArrayList<>();
         userRepository.findAll().forEach(User -> uuidList.add(User.getID()));
-        return uuidList;
+        return ResponseEntity.ok(uuidList);
     }
 
-    // Get User by ID
+    /**
+     * This method gets a user by ID.
+     * @param id Valid ID
+     * @return Request user by ID
+     */
     @GetMapping("/id/{id}")
-    public User getUserById(@PathVariable("id") UUID id) {
-        return userRepository.findById(id)
-                .orElse(null);
+    public ResponseEntity<?> getUserById(@PathVariable("id") UUID id) {
+        if (userRepository.findById(id).isPresent()) {
+            return ResponseEntity.ok(userRepository.findById(id));
+        }
+        return ResponseEntity.badRequest().body("Error! User ID is invalid!");
     }
 
-    // Get Users by Role
+    /**
+     * This method gets all users by role.
+     * @param role role of user
+     * @return List of all users with specified role
+     */
     @GetMapping("/role/{role}")
     public List<User> getUserByRole(@PathVariable("role") String role) {
         List<User> userByRole = new ArrayList<>();
@@ -67,19 +87,34 @@ public class UserController {
         return userByRole;
     }
 
-    // Delete User
+    /**
+     * This method deletes a user by ID
+     * @param id Valid user ID
+     * @return Success or error message
+     */
     @DeleteMapping("/{id}")
-    public void removeUser(@PathVariable("id") UUID id) {
-        userRepository.deleteById(id);
+    public ResponseEntity<?> removeUser(@PathVariable("id") UUID id) {
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.deleteById(id);
+            return ResponseEntity.ok("Success!");
+        }
+        return ResponseEntity.badRequest().body("Error! User ID is invalid!");
     }
 
-    // Update User
+    /**
+     * This method updates a user by user ID
+     * @param id Valid user id
+     * @param user Updated user entity
+     * @return Updated user
+     */
     @PutMapping("/{id}")
-    public User updateUser(@PathVariable("id") UUID id, @Valid @NonNull @RequestBody User user) {
-        User newUser = userRepository.findById(id).orElse(null);
-        assert newUser != null;
-        newUser.setName(user.getName());
-        newUser.setRole(user.getRole());
-        return userRepository.save(newUser);
+    public ResponseEntity<?> updateUser(@PathVariable("id") UUID id, @Valid @NonNull @RequestBody User user) {
+        if (userRepository.findById(id).isPresent()) {
+            User newUser = userRepository.findById(id).get();
+            newUser.setName(user.getName());
+            newUser.setRole(user.getRole());
+            return ResponseEntity.ok(userRepository.save(newUser));
+        }
+        return ResponseEntity.badRequest().body("Error! User ID is invalid!");
     }
 }
